backend/requirements.txt
# Django核心框架
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# 数据库
PyMySQL==1.1.0
redis==5.0.1

# 认证与安全
djangorestframework-simplejwt==5.3.0
cryptography==41.0.7
pycryptodome==3.19.0

# 文件处理
Pillow==10.1.0
python-magic==0.4.27

# 消息队列
celery==5.3.4
kombu==5.3.4
redis==5.0.1

# HTTP请求
requests==2.31.0
urllib3==2.1.0

# 时间处理
python-dateutil==2.8.2
pytz==2023.3

# 配置管理
python-decouple==3.8
django-environ==0.11.2

# 日志
django-extensions==3.2.3

# 开发工具
django-debug-toolbar==4.2.0
ipython==8.17.2

# 测试
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# 代码质量
flake8==6.1.0
black==23.11.0
isort==5.12.0

# 文档
drf-spectacular==0.26.5

# 部署
gunicorn==21.2.0
whitenoise==6.6.0

# 监控
sentry-sdk==1.38.0
